LayoutTests/editing/spelling/spelling-hasspellingmarker-expected.txt

This tests internals.hasSpellingMarker works for differnt type of elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS hasMarked("<textarea id='test' cols='80' rows='10'></textarea>"); is true
PASS hasMarked("<input id='test' type='text'>"); is true
PASS hasMarked("<div id='test' contenteditable='true'></div>"); is true
PASS hasMarked("<div id='test' contentEditable='true' spellcheck='false'></div>"); is false
PASS successfullyParsed is true

TEST COMPLETE
